adapters/llamaindex/pyproject.toml
[project]
name = "llama-index-tools-eydii"
version = "0.2.1"
description = "EYDII + Execute tools for LlamaIndex — verify agent actions and track executions with cryptographic receipts"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Veritera AI", email = "engineering@veritera.ai"}]
keywords = ["veritera", "eydii", "llamaindex", "llama-index", "verification", "guardrail", "ai-safety"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "veritera>=0.2.0",
    "llama-index-core>=0.11.0",
]

[project.urls]
Homepage = "https://id.veritera.ai"
Documentation = "https://id.veritera.ai/docs"
Repository = "https://github.com/veritera-ai/eydii-llamaindex"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["eydii_llamaindex*", "llama_index_tools_eydii*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
